+++ src/types/widget.ts
@@ -4,6 +4,10 @@ export type WidgetType = 'buy' | 'checkout';
 
 export type IntegrationType = 'direct' | 'secure_standalone';
 
+/**
+ * Note: Two factor authentication does not work in an iframe, so the embedded experience should not be used. It will
+ * be removed in a future release.
+ */
 export type Experience = 'embedded' | 'popup' | 'new_tab';
 
 export type Theme = 'light' | 'dark';